build base events layout

+++ app/src/main/java/com/grayzlp/ggithub/data/api/model/event/BaseEvent.java
@@ -19,5 +19,6 @@ public class BaseEvent {
     public Actor actor;
     public Organization org;
     public Date create_at;
-    public int id;
+    public long id;
+
 }
